+++ glade_forms/lib/src/core/error/convert_error.dart
@@ -1,28 +1,31 @@
 import 'package:equatable/equatable.dart';
-import 'package:glade_forms/src/core/error/glade_input_error.dart';
+import 'package:glade_forms/src/core/error/glade_input_validation.dart';
+import 'package:glade_forms/src/core/error/validation_severity.dart';
 
 /// Before validation when converer from string to prpoer type failed.
 typedef OnConvertError = String Function(String? rawInput, {Object? key});
 
-class ConvertError<T> extends GladeInputError<T> with EquatableMixin implements Exception {
-  // ignore: prefer-correct-callback-field-name, more suitable name
-  final OnConvertError devError;
+class ConvertError<T> extends GladeInputValidation<T> with EquatableMixin implements Exception {
+  final OnConvertError onConvertErrorMessage;
 
   final String? input;
 
   // ignore: no-object-declaration, error can be anything (but typically it is string)
-  final Object? _convertError;
+  final Object _convertError;
 
   @override
-  List<Object?> get props => [input, devError, error, key, _convertError];
+  List<Object?> get props => [input, onConvertErrorMessage, result, key, _convertError];
 
   String get targetType => T.runtimeType.toString();
 
-  String get devErrorMessage => devError(input, key: key);
+  String get devMessage => onConvertErrorMessage(input, key: key);
 
   @override
   // ignore: no-object-declaration, error can be any object.
-  Object? get error => _convertError;
+  Object get result => _convertError;
+
+  @override
+  ValidationSeverity get severity => ValidationSeverity.error;
 
   ConvertError({
     required Object error,
@@ -30,8 +33,8 @@ class ConvertError<T> extends GladeInputError<T> with EquatableMixin implements 
     super.key,
     OnConvertError? formatError,
   })  : _convertError = error,
-        devError = formatError ?? ((rawValue, {key}) => 'Conversion error: $error');
+        onConvertErrorMessage = formatError ?? ((rawValue, {key}) => 'Conversion error: $error');
 
   @override
-  String toString() => devError(input, key: key);
+  String toString() => onConvertErrorMessage(input, key: key);
 }