+++ glade_forms/lib/src/core/error/glade_validations_keys.dart
@@ -1,4 +1,4 @@
-abstract final class GladeErrorKeys {
+abstract final class GladeValidationsKeys {
   static const String conversionError = 'value-cant-be-converted';
   static const String stringEmpty = 'string-empty-error';
   static const String stringNotUrl = 'string-not-url';