+++ glade_forms/lib/src/core/input/glade_string_input.dart
@@ -1,4 +1,4 @@
-import 'package:glade_forms/src/core/error/glade_error_keys.dart';
+import 'package:glade_forms/src/core/error/glade_validations_keys.dart';
 import 'package:glade_forms/src/core/input/glade_input.dart';
 import 'package:glade_forms/src/validator/part/satisfy_predicate_part.dart';
 import 'package:glade_forms/src/validator/specialized/string_validator.dart';
@@ -17,7 +17,7 @@ class GladeStringInput extends GladeInput<String> {
     String? initialValue,
     StringValidatorFactory? validator,
     super.isPure,
-    super.translateError,
+    super.validationTranslate,
     super.valueComparator,
     super.stringToValueConverter,
     super.dependencies,
@@ -26,7 +26,7 @@ class GladeStringInput extends GladeInput<String> {
     super.textEditingController,
     super.useTextEditingController = true,
     super.valueTransform,
-    super.defaultTranslations,
+    super.defaultValidationTranslations,
     super.trackUnchanged = true,
     bool isRequired = true,
   }) : super.internalCreate(
@@ -39,7 +39,7 @@ class GladeStringInput extends GladeInput<String> {
 
   /// Maximum length of the string if `maxLength()` standard validator is used.
   int? getMaxLength({Object? key}) {
-    final validator = validatorInstance.tryFindValidatorPart(key ?? GladeErrorKeys.stringMaxLength);
+    final validator = validatorInstance.tryFindValidatorPart(key ?? GladeValidationsKeys.stringMaxLength);
 
     if (validator == null) return null;