+++ glade_forms/lib/src/validator/validator_error/value_error.dart
@@ -1,5 +1,0 @@
-import 'package:glade_forms/src/validator/validator_error/glade_validator_error.dart';
-
-class ValueError<T> extends GladeValidatorError<T> {
-  ValueError({required super.value, required super.devError, required super.key});
-}